+++ app/models/conversion/task_list.rb
@@ -10,8 +10,7 @@ class Conversion::TaskList < ::BaseTaskList
           Conversion::Task::CompleteNotificationOfChangeTaskForm,
           Conversion::Task::ConversionGrantTaskForm,
           Conversion::Task::SponsoredSupportGrantTaskForm,
-          Conversion::Task::AcademyDetailsTaskForm,
-          Conversion::Task::FundingAgreementContactTaskForm
+          Conversion::Task::AcademyDetailsTaskForm
         ]
       },
       {