+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased][unreleased]
 
+### Changed
+
+- Remove the Funding agreement letters contact task from the Conversion project
+  task list
+
 ## [Release 39][release-39]
 
 ### Added